Restructure and Renaming
- Moved some files to config folder
- Move config folder to root
- Renamed some files to match convention
- Added a new favicon to more accurately reflect the app's purpose
