+++ config/eslint.config.js
