+++ src/components/App.tsx
@@ -1,14 +1,14 @@
 import '../styles/app.css'
 import { useState, useEffect, useRef } from 'react'
-import { onAuthChange } from '../services/authService'
-import { fetchDrawings, saveDrawing, Drawing } from '../services/drawingsService'
-import { auth } from '../../firebaseConfig'
+import { onAuthChange } from '../services/AuthService'
+import { fetchDrawings, saveDrawing, Drawing } from '../services/DrawingsService'
+import { auth } from '../../config/firebaseConfig'
 import Canvas, { CanvasHandle, gridWidth, gridHeight } from '../components/Canvas'
 import AuthModal from '../components/AuthModal'
 import ColourPicker from '../components/ColourPicker'
 import SidePanel from '../components/SidePanel'
 import ConfirmationPopup from '../components/ConfirmationPopup';
-import { LOCAL_STORAGE_KEYS } from '../constants'
+import { LOCAL_STORAGE_KEYS } from '../utils/constants'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faFloppyDisk, faDownload, faFile, faUpload } from '@fortawesome/free-solid-svg-icons'
 import { useClientStateStore } from '../services/ClientStateService'
