+++ tailwind.config.js
@@ -16,5 +16,4 @@ module.exports = {
     extend: {},
   },
   plugins: [],
-};
-
+}
